refactor: gameclock
