+++ boolebot/src/SASS/Partials/_mixins.scss
@@ -37,7 +37,7 @@
         }
     
         
-        input {
+        input[type="text"] {
             margin-left: 10px;
         }
